backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.28.1
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.2
pydantic==2.8.2
python-dotenv==1.0.1
endee>=0.1.0
sentence-transformers>=2.7.0
